Move AuthToken and UserCredentials data classes under java/data/user folder
