+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/data/user/UserCredentials.kt
@@ -1,4 +1,4 @@
-package com.bounswe2020group3.paperlayer.profile.data.data
+package com.bounswe2020group3.paperlayer.data.user
 
 import com.squareup.moshi.Json
 data class UserCredentials (
