+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/invite/InviteModel.kt
@@ -3,7 +3,7 @@ package com.bounswe2020group3.paperlayer.invite
 import com.bounswe2020group3.paperlayer.invite.data.InviteRequest
 import com.bounswe2020group3.paperlayer.invite.data.InviteResponse
 import com.bounswe2020group3.paperlayer.invite.data.CollaborationInvite
-import com.bounswe2020group3.paperlayer.profile.data.data.AuthToken
+import com.bounswe2020group3.paperlayer.data.user.AuthToken
 import com.bounswe2020group3.paperlayer.data.user.User
 import com.bounswe2020group3.paperlayer.util.Session
 import io.reactivex.Observable
