+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/login/LoginPresenter.kt
@@ -3,7 +3,7 @@ package com.bounswe2020group3.paperlayer.login
 
 import androidx.navigation.Navigation
 import com.bounswe2020group3.paperlayer.R
-import com.bounswe2020group3.paperlayer.profile.data.data.UserCredentials
+import com.bounswe2020group3.paperlayer.data.user.UserCredentials
 import com.bounswe2020group3.paperlayer.mvp.BasePresenter
 import io.reactivex.disposables.CompositeDisposable
 import javax.inject.Inject
