+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/ProfileModel.kt
@@ -1,6 +1,6 @@
 package com.bounswe2020group3.paperlayer.profile
 
-import com.bounswe2020group3.paperlayer.profile.data.data.AuthToken
+import com.bounswe2020group3.paperlayer.data.user.AuthToken
 import com.bounswe2020group3.paperlayer.network.RetrofitProvider
 import com.bounswe2020group3.paperlayer.data.user.Profile
 import com.bounswe2020group3.paperlayer.data.user.User
