+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/project/ProjectsContract.kt
@@ -1,6 +1,6 @@
 package com.bounswe2020group3.paperlayer.project
 
-import com.bounswe2020group3.paperlayer.profile.data.data.AuthToken
+import com.bounswe2020group3.paperlayer.data.user.AuthToken
 import com.bounswe2020group3.paperlayer.mvp.Mvp
 import com.bounswe2020group3.paperlayer.project.data.Project
 import com.bounswe2020group3.paperlayer.project.data.ProjectShort
